LPS-79031 Move GroupWrapper out of architect package
